Remove interface and expose class to out of lib
