+++ core/src/main/kotlin/kotlet/MethodNotFoundException.kt
@@ -3,4 +3,4 @@ package kotlet
 /**
  * Exception thrown when a method is not found. For internal use only.
  */
-internal class MethodNotFoundException : Exception()
+internal class MethodNotFoundException : RuntimeException()
