+++ core/src/main/kotlin/kotlet/RoutingConfigurationException.kt
@@ -1,3 +1,6 @@
 package kotlet
 
+/**
+ * Exception thrown when there is an error in the routing configuration.
+ */
 class RoutingConfigurationException(message: String) : RuntimeException(message)
